Add EAR packaging
