+++ openmrs-keycloak-userstore/src/main/resources/META-INF/services/org.keycloak.credential.CredentialInputValidator
