+++ openmrs-keycloak-userstore/src/main/resources/META-INF/services/org.keycloak.storage.UserStorageProviderFactory
@@ -1,0 +1,1 @@
+org.openmrs.contrib.keycloak.userstore.provider.OpenmrsAuthenticatorProviderFactory
