+++ openmrs-keycloak-userstore/src/test/resources/data.sql
