--- src/main/resources/META-INF/services/org.keycloak.storage.UserStorageProviderFactory
@@ -1,1 +1,0 @@
-org.openmrs.keycloak.provider.OpenmrsAuthenticatorProviderFactory
